# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  helperdto.BurnCredentialVCParamDTO:
    description: Data Transfer Object for burning a verifiable credential
    properties:
      did:
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        type: string
      option:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      token_id:
        example: "1"
        type: string
    required:
    - did
    - option
    - token_id
    type: object
  helperdto.ClaimCredentialVCParamDTO:
    description: Data Transfer Object for claiming a verifiable credential
    properties:
      option:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      token_id:
        example: "1"
        type: string
    required:
    - option
    - token_id
    type: object
...
  helperdto.EncodeBurnCredentialDTO:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
      vc_param:
        $ref: '#/definitions/helperdto.BurnCredentialVCParamDTO'
    required:
    - address
    - vc_param
    type: object
  helperdto.EncodeClaimCredentialDTO:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
      vc_param:
        $ref: '#/definitions/helperdto.ClaimCredentialVCParamDTO'
    required:
    - address
    - vc_param
    type: object
...
  helperdto.EncodeIssueCredentialDTO:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
      vc_param:
        $ref: '#/definitions/helperdto.IssueCredentialVCParamDTO'
    required:
    - address
    - vc_param
    type: object
...
  helperdto.EncodeRevokeCredentialDTO:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
      vc_param:
        $ref: '#/definitions/helperdto.RevokeCredentialVCParamDTO'
    required:
    - address
    - vc_param
    type: object
...
  helperdto.EncodeUpdateMetadataDTO:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
      vc_param:
        $ref: '#/definitions/helperdto.UpdateMetadataVCParamDTO'
    required:
    - address
    - vc_param
    type: object
...
  helperdto.IssueCredentialVCParamDTO:
    description: Data Transfer Object for issuing a verifiable credential
    properties:
      did_recipient:
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        type: string
      expired_at:
        example: "1735689600"
        type: string
      offchain_hash:
        example: QmXnnyufdzAWL5CqZ2RnSNgPbvCc1ALT73s6epPrRnZ1Xy
        type: string
      option:
        enum:
        - 0
        - 1
        example: 0
        type: integer
    required:
    - did_recipient
    - expired_at
    - offchain_hash
    - option
    type: object
...
  helperdto.RevokeCredentialVCParamDTO:
    description: Data Transfer Object for revoking a verifiable credential
    properties:
      option:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      token_id:
        example: "1"
        type: string
    required:
    - option
    - token_id
    type: object
...
  helperdto.StringToHex32DTO:
    properties:
      input:
        example: example_string
        type: string
    required:
    - input
    type: object
...
  helperdto.UpdateMetadataVCParamDTO:
    description: Data Transfer Object for updating credential metadata
    properties:
      expired_at:
        example: "1735689600"
        type: string
      is_valid:
        example: true
        type: boolean
      offchain_hash:
        example: QmXnnyufdzAWL5CqZ2RnSNgPbvCc1ALT73s6epPrRnZ1Xy
        type: string
      option:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      token_id:
        example: "1"
        type: string
    required:
    - expired_at
    - offchain_hash
    - option
    - token_id
    type: object
  helperhandler.HelperResponse:
    properties:
      data:
        example: 0x...
        type: string
      details:
        $ref: '#/definitions/response.Details'
      nonce:
        example: "1"
        type: string
      target:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
    type: object
...
  response.Details:
    properties:
      service:
        example: 0
        type: integer
      tx_type:
        example: 10
        type: integer
    type: object
  response.Meta:
    properties:
      timestamp:
        type: string
      version:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: DID Backend API
  title: DID API
  version: "1.0"
paths:
...
  /helper/encode-burn-credential-param:
    post:
      consumes:
      - application/json
      description: Encodes parameters for burning/destroying a verifiable credential
      parameters:
      - description: Burn Credential parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/helperdto.EncodeBurnCredentialDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/helperhandler.HelperResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Encode Burn Credential parameters
      tags:
      - Helper - DID VC
  /helper/encode-claim-credential-param:
    post:
      consumes:
      - application/json
      description: Encodes parameters for claiming an unclaimed verifiable credential
      parameters:
      - description: Claim Credential parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/helperdto.EncodeClaimCredentialDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/helperhandler.HelperResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Encode Claim Credential parameters
      tags:
      - Helper - DID VC
...
  /helper/encode-issue-credential-param:
    post:
      consumes:
      - application/json
      description: Encodes parameters for issuing a new verifiable credential
      parameters:
      - description: Issue Credential parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/helperdto.EncodeIssueCredentialDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/helperhandler.HelperResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Encode Issue Credential parameters
      tags:
      - Helper - DID VC
...
  /helper/encode-revoke-credential-param:
    post:
      consumes:
      - application/json
      description: Encodes parameters for revoking a verifiable credential
      parameters:
      - description: Revoke Credential parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/helperdto.EncodeRevokeCredentialDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/helperhandler.HelperResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Encode Revoke Credential parameters
      tags:
      - Helper - DID VC
...
  /helper/encode-update-metadata-param:
    post:
      consumes:
      - application/json
      description: Encodes parameters for updating credential metadata
      parameters:
      - description: Update Metadata parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/helperdto.EncodeUpdateMetadataDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/helperhandler.HelperResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Encode Update Metadata parameters
      tags:
      - Helper - DID VC
  /helper/string-2-byte32:
    post:
      consumes:
      - application/json
      description: |-
        Converts a string input to a 32-byte array (byte32). This is a utility endpoint commonly used for encoding data that needs to fit into a single bytes32 storage slot in smart contracts.

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true for HTTP 200
        - `data` (string): The hex-encoded 32-byte array result
        - `meta` (object): Contains timestamp and API version

        ## Common Use Cases
        - Encoding strings for smart contract storage
        - Preparing data for keccak256 hashing
        - Converting identifiers to fixed-size byte arrays

        ## Important Notes
        - Input strings longer than 32 bytes will be truncated
        - Shorter strings will be padded with zeros
        - The result is returned as a hex string with 0x prefix
      parameters:
      - description: String input to convert to byte32
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/helperdto.StringToHex32DTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully converted string to byte32
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid request body - malformed JSON or missing required fields
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - conversion failed
          schema:
            $ref: '#/definitions/response.Response'
      summary: Convert String to Byte32
      tags:
      - helper General
swagger: "2.0"